+++ app/src/main/java/jp/example/tanmen/Adapter/BottomNavigationPagerAdapter.kt
@@ -1,10 +1,10 @@
-package jp.example.tanmen.View.Adapter
+package jp.example.tanmen.Adapter
 
 import androidx.fragment.app.Fragment
 import androidx.viewpager2.adapter.FragmentStateAdapter
-import jp.example.tanmen.View.Fragment.HomeFragment
-import jp.example.tanmen.View.Fragment.MainFragment
-import jp.example.tanmen.View.Fragment.ShuffleFragment
+import jp.example.tanmen.view.Fragment.HomeFragment
+import jp.example.tanmen.view.Fragment.MainFragment
+import jp.example.tanmen.view.Fragment.ShuffleFragment
 
 class BottomNavigationPagerAdapter(fm: MainFragment) : FragmentStateAdapter(fm) {
     override fun getItemCount(): Int {
