+++ app/src/main/res/layout/fragment_main.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:id="@+id/container"
-    tools:context=".View.Fragment.MainFragment">
+    tools:context=".view.Fragment.MainFragment">
 
     <androidx.viewpager2.widget.ViewPager2
         android:id="@+id/viewPager"
